+++ builds/crx/script.js
@@ -10245,7 +10245,7 @@
 
   Redirect = {
     init: function() {
-      var archive, archives, boardID, boards, data, files, id, name, o, record, software, type, _i, _j, _len, _len1, _ref, _ref1, _ref2;
+      var archive, archives, boardID, boards, data, files, id, name, o, record, software, type, _i, _j, _len, _len1, _ref, _ref1;
       o = {
         thread: {},
         post: {},
@@ -10254,13 +10254,9 @@
       archives = {};
       _ref = Redirect.archives;
       for (_i = 0, _len = _ref.length; _i < _len; _i++) {
-        _ref1 = _ref[_i], name = _ref1.name, boards = _ref1.boards, files = _ref1.files, data = _ref1.data;
-        archives[name] = {
-          boards: boards,
-          files: files,
-          data: data
-        };
-        software = data.software;
+        data = _ref[_i];
+        name = data.name, boards = data.boards, files = data.files, software = data.software;
+        archives[name] = data;
         for (_j = 0, _len1 = boards.length; _j < _len1; _j++) {
           boardID = boards[_j];
           if (!(boardID in o.thread)) {
@@ -10274,9 +10270,9 @@
           }
         }
       }
-      _ref2 = Conf['selectedArchives'];
-      for (boardID in _ref2) {
-        record = _ref2[boardID];
+      _ref1 = Conf['selectedArchives'];
+      for (boardID in _ref1) {
+        record = _ref1[boardID];
         for (type in record) {
           id = record[type];
           if (!((archive = archives[id]))) {
@@ -10286,141 +10282,12 @@
           if (__indexOf.call(boards, boardID) < 0) {
             continue;
           }
-          o[type][boardID] = archive.data;
+          o[type][boardID] = archive;
         }
       }
       return Redirect.data = o;
     },
-    archives: [
-      {
-        name: "Foolz",
-        boards: ["a", "biz", "co", "diy", "gd", "jp", "m", "sci", "sp", "tg", "tv", "v", "vg", "vp", "vr", "wsg"],
-        files: ["a", "biz", "diy", "gd", "jp", "m", "sci", "tg", "vg", "vp", "vr", "wsg"],
-        data: {
-          domain: "archive.foolz.us",
-          http: false,
-          https: true,
-          software: "foolfuuka"
-        }
-      }, {
-        name: "NSFW Foolz",
-        boards: ["u"],
-        files: ["u"],
-        data: {
-          domain: "nsfw.foolz.us",
-          http: false,
-          https: true,
-          software: "foolfuuka"
-        }
-      }, {
-        name: "The Dark Cave",
-        boards: ["c", "int", "out", "po"],
-        files: ["c", "po"],
-        data: {
-          domain: "archive.thedarkcave.org",
-          http: true,
-          https: true,
-          software: "foolfuuka"
-        }
-      }, {
-        name: "4plebs",
-        boards: ["adv", "hr", "o", "pol", "s4s", "tg", "tv", "x"],
-        files: ["adv", "hr", "o", "pol", "s4s", "tg", "tv", "x"],
-        data: {
-          domain: "archive.4plebs.org",
-          http: true,
-          https: true,
-          software: "foolfuuka"
-        }
-      }, {
-        name: "Nyafuu",
-        boards: ["c", "e", "w", "wg"],
-        files: ["c", "e", "w", "wg"],
-        data: {
-          domain: "archive.nyafuu.org",
-          http: true,
-          https: true,
-          software: "foolfuuka"
-        }
-      }, {
-        name: "Love is Over",
-        boards: ["d", "i"],
-        files: ["d", "i"],
-        data: {
-          domain: "loveisover.me",
-          http: true,
-          https: true,
-          software: "foolfuuka"
-        }
-      }, {
-        name: "Rebecca Black Tech",
-        boards: ["cgl", "g", "mu", "w"],
-        files: ["cgl", "g", "mu", "w"],
-        data: {
-          domain: "archive.rebeccablacktech.com",
-          http: true,
-          https: true,
-          software: "fuuka"
-        }
-      }, {
-        name: "Heinessen",
-        boards: ["an", "fit", "k", "mlp", "r9k", "toy"],
-        files: ["an", "fit", "k", "r9k", "toy"],
-        data: {
-          domain: "archive.heinessen.com",
-          http: true,
-          software: "fuuka"
-        }
-      }, {
-        name: "warosu",
-        boards: ["3", "biz", "cgl", "ck", "diy", "fa", "g", "ic", "jp", "lit", "sci", "tg", "vr"],
-        files: ["3", "biz", "cgl", "ck", "diy", "fa", "ic", "jp", "lit", "sci", "tg", "vr"],
-        data: {
-          domain: "fuuka.warosu.org",
-          https: true,
-          software: "fuuka"
-        }
-      }, {
-        name: "fgts",
-        boards: ["r", "soc"],
-        files: ["r", "soc"],
-        data: {
-          domain: "fgst.eu",
-          http: true,
-          https: true,
-          software: "foolfuuka"
-        }
-      }, {
-        name: "maware",
-        boards: ["t"],
-        files: ["t"],
-        data: {
-          domain: "archive.mawa.re",
-          http: true,
-          software: "foolfuuka"
-        }
-      }, {
-        name: "InstallGentoo",
-        boards: ["g", "t"],
-        files: ["g", "t"],
-        data: {
-          domain: "chan.installgentoo.com",
-          http: true,
-          software: "foolfuuka"
-        }
-      }, {
-        name: "Foolz Beta",
-        boards: ["a", "biz", "co", "d", "diy", "gd", "jp", "m", "mlp", "s4s", "sci", "sp", "tg", "tv", "u", "v", "vg", "vp", "vr", "wsg"],
-        files: ["a", "biz", "d", "diy", "gd", "jp", "m", "s4s", "sci", "tg", "u", "vg", "vp", "vr", "wsg"],
-        data: {
-          domain: "beta.foolz.us",
-          http: true,
-          https: true,
-          withCredentials: true,
-          software: "foolfuuka"
-        }
-      }
-    ],
+    archives: [{"uid":0,"name":"Foolz","domain":"archive.foolz.us","http":true,"https":true,"software":"foolfuuka","boards":["a","biz","co","diy","gd","jp","m","sci","sp","tg","tv","v","vg","vp","vr","wsg"],"files":["a","biz","gd","diy","jp","m","sci","tg","vg","vp","vr","wsg"]},{"uid":1,"name":"NSFW Foolz","domain":"nsfw.foolz.us","http":true,"https":true,"software":"foolfuuka","boards":["u"],"files":["u"]},{"uid":2,"name":"The Dark Cave","domain":"archive.thedarkcave.org","http":true,"https":true,"software":"foolfuuka","boards":["c","int","out","po"],"files":["c","po"]},{"uid":3,"name":"4plebs Archive","domain":"archive.4plebs.org","http":true,"https":true,"software":"foolfuuka","boards":["adv","hr","o","pol","s4s","tg","trv","tv","x"],"files":["adv","hr","o","pol","s4s","tg","trv","tv","x"]},{"uid":18,"name":"4plebs Flash Archive","domain":"flash.4plebs.org","http":true,"https":true,"software":"foolfuuka","boards":["f"],"files":["f"]},{"uid":4,"name":"Nyafuu","domain":"archive.nyafuu.org","http":true,"https":true,"software":"foolfuuka","boards":["c","e","w","wg"],"files":["c","e","w","wg"]},{"uid":5,"name":"Love is Over","domain":"loveisover.me","http":true,"https":true,"software":"foolfuuka","boards":["d","i"],"files":["d","i"]},{"uid":8,"name":"Rebecca Black Tech","domain":"archive.rebeccablacktech.com","http":true,"https":true,"software":"fuuka","boards":["cgl","g","mu","w"],"files":["cgl","g","mu","w"]},{"uid":9,"name":"Heinessen","domain":"archive.heinessen.com","http":true,"https":false,"software":"fuuka","boards":["an","fit","k","mlp","r9k","toy"],"files":["an","fit","k","r9k","toy"]},{"uid":10,"name":"warosu","domain":"fuuka.warosu.org","http":false,"https":true,"software":"fuuka","boards":["3","biz","cgl","ck","diy","fa","g","ic","jp","lit","sci","tg","vr"],"files":["3","biz","cgl","ck","diy","fa","ic","jp","lit","sci","tg","vr"]},{"uid":15,"name":"fgts","domain":"fgts.eu","http":true,"https":true,"software":"foolfuuka","boards":["cm","hm","r","soc","y"],"files":["cm","hm","r","soc","y"]},{"uid":16,"name":"maware","domain":"archive.mawa.re","http":true,"https":false,"software":"foolfuuka","boards":["t"],"files":["t"]},{"uid":17,"name":"installgentoo.com","domain":"chan.installgentoo.com","http":true,"https":false,"software":"foolfuuka","boards":["g","t"],"files":["g","t"]},{"uid":13,"name":"Foolz Beta","domain":"beta.foolz.us","http":true,"https":true,"withCredentials":true,"software":"foolfuuka","boards":["a","biz","co","d","diy","gd","jp","m","s4s","sci","sp","tg","tv","u","v","vg","vp","vr","wsg"],"files":["a","biz","d","diy","gd","jp","m","s4s","sci","tg","u","vg","vp","vr","wsg"]}],
     to: function(dest, data) {
       var archive;
       archive = (dest === 'search' ? Redirect.data.thread : Redirect.data[dest])[data.boardID];
@@ -12905,7 +12772,7 @@
       return $.on(ta, 'change', $.cb.value);
     },
     advanced: function(section) {
-      var archBoards, boardID, boardOptions, boardSelect, boards, data, event, files, input, inputs, item, items, name, o, row, rows, ta, table, _i, _j, _k, _l, _len, _len1, _len2, _len3, _len4, _m, _ref, _ref1, _ref2, _ref3, _ref4;
+      var archBoards, boardID, boardOptions, boardSelect, boards, event, files, input, inputs, item, items, name, o, row, rows, software, ta, table, _i, _j, _k, _l, _len, _len1, _len2, _len3, _len4, _m, _ref, _ref1, _ref2, _ref3, _ref4;
       section.innerHTML = "<fieldset><legend>Archiver</legend><div class=\"warning\" " + (Conf['404 Redirect'] ? 'hidden' : '') + "><code>404 Redirect</code> is disabled.</div><select id='archive-board-select'></select><table id='archive-table'><thead><th>Thread redirection</th><th>Post fetching</th><th>File redirection</th></thead><tbody></tbody></table><span class=note>Disabled selections indicate that only one archive is available for that board and redirection type.</span></fieldset><fieldset><legend>Custom Board Navigation</span></legend><div><textarea name=boardnav class=field spellcheck=false></textarea></div><span class=note>New lines will be converted into spaces.</span><br><br><div class=note>In the following examples for /g/, <code>g</code> can be changed to a different board ID (<code>a</code>, <code>b</code>, etc...), the current board (<code>current</code>), or the Twitter link (<code>@</code>).</div><div>Board link: <code>g</code></div><div>Title link: <code>g-title</code></div><div>Board link (Replace with title when on that board): <code>g-replace</code></div><div>Full text link: <code>g-full</code></div><div>Custom text link: <code>g-text:\"Install Gentoo\"</code></div><div>Index-only link: <code>g-index</code></div><div>Catalog-only link: <code>g-catalog</code></div><div>External link: <code>external-text:\"Google\",\"http://www.google.com\"</code></div><div>Combinations are possible: <code>g-index-text:\"Technology Index\"</code></div><div>Full board list toggle: <code>toggle-all</code></div><br><div class=note><code>[ toggle-all ] [current-title] [g-title / a-title / jp-title] [x / wsg / h] [t-text:\"Piracy\"]</code><br>will give you<br><code>[ + ] [Technology] [Technology / Anime & Manga / Otaku Culture] [x / wsg / h] [Piracy]</code><br>if you are on /g/.</div></fieldset><fieldset><legend>Time Formatting <span class=warning " + (Conf['Time Formatting'] ? 'hidden' : '') + ">is disabled.</span></legend><div><input name=time class=field spellcheck=false>: <span class=time-preview></span></div><div>Supported <a href=//en.wikipedia.org/wiki/Date_%28Unix%29#Formatting>format specifiers</a>:</div><div>Day: <code>%a</code>, <code>%A</code>, <code>%d</code>, <code>%e</code></div><div>Month: <code>%m</code>, <code>%b</code>, <code>%B</code></div><div>Year: <code>%y</code>, <code>%Y</code></div><div>Hour: <code>%k</code>, <code>%H</code>, <code>%l</code>, <code>%I</code>, <code>%p</code>, <code>%P</code></div><div>Minute: <code>%M</code></div><div>Second: <code>%S</code></div></fieldset><fieldset><legend>Quote Backlinks formatting <span class=warning " + (Conf['Quote Backlinks'] ? 'hidden' : '') + ">is disabled.</span></legend><div><input name=backlink class=field spellcheck=false>: <span class=backlink-preview></span></div></fieldset><fieldset><legend>File Info Formatting <span class=warning " + (Conf['File Info Formatting'] ? 'hidden' : '') + ">is disabled.</span></legend><div><input name=fileInfo class=field spellcheck=false>: <span class='fileText file-info-preview'></span></div><div>Link: <code>%l</code> (truncated), <code>%L</code> (untruncated), <code>%T</code> (Unix timestamp)</div><div>Original file name: <code>%n</code> (truncated), <code>%N</code> (untruncated), <code>%t</code> (Unix timestamp)</div><div>Spoiler indicator: <code>%p</code></div><div>Size: <code>%B</code> (Bytes), <code>%K</code> (KB), <code>%M</code> (MB), <code>%s</code> (4chan default)</div><div>Resolution: <code>%r</code> (Displays 'PDF' for PDF files)</div></fieldset><fieldset><legend>Quick Reply Personas</legend><textarea class=personafield name=\"QR.personas\" class=\"field\" spellcheck=\"false\"></textarea><p>One item per line.<br>Items will be added in the relevant input's auto-completion list.<br>Password items will always be used, since there is no password input.<br>Lines starting with a <code>#</code> will be ignored.</p><ul>You can use these settings with each item, separate them with semicolons:<li>Possible items are: <code>name</code>, <code>email</code>, <code>subject</code> and <code>password</code>.</li><li>Wrap values of items with quotes, like this: <code>email:\"sage\"</code>.</li><li>Force values as defaults with the <code>always</code> keyword, for example: <code>email:\"sage\";always</code>.</li><li>Select specific boards for an item, separated with commas, for example: <code>email:\"sage\";boards:jp;always</code>.</li></ul></fieldset><fieldset><legend>Unread Favicon <span class=warning " + (Conf['Unread Favicon'] ? 'hidden' : '') + ">is disabled.</span></legend><select name=favicon><option value=ferongr>ferongr</option><option value=xat->xat-</option><option value=4chanJS>4chanJS</option><option value=Mayhem>Mayhem</option><option value=Original>Original</option><option value=Metro>Metro</option></select><span class=favicon-preview></span></fieldset><fieldset><legend>Emoji <span class=warning " + (Conf['Emoji'] ? 'hidden' : '') + ">is disabled.</span></legend><div>Sage Icon: <select name=sageEmoji><option value=\"4chan SS\">4chan SS</option><option value=\"appchan\">appchan</option></select><span class=sage-icon-preview></span></div><div>Position: <select name=emojiPos><option value=\"before\">Before</option><option value=\"after\">After</option></select></div></fieldset><fieldset><legend>Thread Updater <span class=warning " + (Conf['Thread Updater'] ? 'hidden' : '') + ">is disabled.</span></legend><div>Interval: <input type=number name=Interval class=field min=1 value=" + Conf['Interval'] + "></div></fieldset><fieldset><legend><label><input type=checkbox name='Custom CSS' " + (Conf['Custom CSS'] ? 'checked' : '') + "> Custom CSS</label></legend><button id=apply-css>Apply CSS</button><textarea name=usercss class=field spellcheck=false " + (Conf['Custom CSS'] ? '' : 'disabled') + "></textarea></fieldset>";
       items = {};
       inputs = {};
@@ -12945,7 +12812,7 @@
       archBoards = {};
       _ref1 = Redirect.archives;
       for (_j = 0, _len1 = _ref1.length; _j < _len1; _j++) {
-        _ref2 = _ref1[_j], name = _ref2.name, boards = _ref2.boards, files = _ref2.files, data = _ref2.data;
+        _ref2 = _ref1[_j], name = _ref2.name, boards = _ref2.boards, files = _ref2.files, software = _ref2.software;
         for (_k = 0, _len2 = boards.length; _k < _len2; _k++) {
           boardID = boards[_k];
           o = archBoards[boardID] || (archBoards[boardID] = {
@@ -12954,7 +12821,7 @@
             file: []
           });
           o.thread.push(name);
-          if (data.software === 'foolfuuka') {
+          if (software === 'foolfuuka') {
             o.post.push(name);
           }
           if (__indexOf.call(files, boardID) >= 0) {
@@ -12993,7 +12860,7 @@
         return $("tbody > ." + this.value, table).hidden = false;
       });
       $.get('selectedArchives', Conf['selectedArchives'], function(_arg) {
-        var option, selectedArchives, type;
+        var data, option, selectedArchives, type;
         selectedArchives = _arg.selectedArchives;
         for (boardID in selectedArchives) {
           data = selectedArchives[boardID];